api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
python-dotenv
pydantic
email-validator
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.23
alembic==1.13.1
python-multipart